requirements.txt
pyserial
structlog==23.1.0
asyncio-paho==0.6.0
